Rename MarkdownTeXParser to MarkdownMathParser
