+++ packages/app-mobile/components/NoteEditor/CodeMirrorTheme.ts
@@ -8,7 +8,7 @@ import { tags } from '@lezer/highlight';
 import { EditorView } from '@codemirror/view';
 import { Extension } from '@codemirror/state';
 
-import { inlineMathTag, mathTag } from './MarkdownTeXParser';
+import { inlineMathTag, mathTag } from './MarkdownMathParser';
 
 // For an example on how to customize the theme, see:
 //
